backend/requirements.txt
# Core Django
Django>=6.0,<7.0
djangorestframework>=3.14.0

# Authentication
djangorestframework-simplejwt>=5.3.0

# CORS
django-cors-headers>=4.3.0

# Database
mysqlclient>=2.2.0

# Cache
django-redis>=5.4.0
redis>=5.0.0

# Image handling
Pillow>=10.0.0
easy-thumbnails>=2.8.5
django-cleanup>=8.0.0

# Testing & Development
Faker>=20.0.0
pytest>=8.0.0
pytest-django>=4.8.0
pytest-cov>=5.0.0
coverage>=7.0.0

# HTTP requests
requests>=2.31.0

# Environment variables
python-decouple>=3.8

# ASGI/WSGI
gunicorn>=21.2.0
